requirements.txt
asgiref==3.8.1
Django==5.0.6
django-crispy-forms==2.2
gunicorn==25.0.3
packaging==26.0
pillow==10.4.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.0
typing_extensions==4.12.2
